+++ app/class-core.php
@@ -179,6 +179,7 @@ class Core {
 	 * @see Integrations\Wpml
 	 * @see Integrations\Shortpixel
 	 * @see Integrations\JS_Composer
+	 * @see Integrations\Flatsome
 	 */
 	private function init_integrations() {
 		$loader = Loader::get_instance();
@@ -191,6 +192,7 @@ class Core {
 		$loader->integration( 'shortpixel' );
 		$loader->integration( 'elementor' );
 		$loader->integration( 'js-composer' );
+		$loader->integration( 'flatsome' );
 	}
 
 	/**