+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 Added:
 * Integration with Elementor Pro Gallery
 
+Changed:
+* Improve image AI modules
+
+Fixed:
+* Only allow generating image alt text for supported formats (JPEG, PNG, GIF, BMP)
+
 = 1.9.1 - 23.04.2024 =
 
 Added: