+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Added:
 * Integration with Elementor Pro Gallery
+* cf_images_upload_host filter to adjust the image host ID
 
 Changed:
 * Improve image AI modules