+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Added:
 * Integration with WPBakery page builder image galleries
 * Integration with Elementor Pro Gallery
+* Integration with Flatsome theme gallery
 * cf_images_upload_host filter to adjust the image host ID
 
 Changed: